--- students/21K0597/21K0597-p02/src/main/java/ru/mirea/practice/Task10/Main.java
@@ -1,11 +1,0 @@
-package ru.mirea.practice.Task10;
-
-import java.util.Scanner;
-
-abstract class Main {
-    public static void main(String[] args) {
-        try (Scanner in = new Scanner(System.in)) {
-            System.out.println(in.nextLine().split("\\s+").length);
-        }
-    }
-}
