--- students/21K0597/21K0597-p02/src/main/java/ru/mirea/practice/Task3/Main.java
@@ -1,10 +1,0 @@
-package ru.mirea.practice.Task3;
-
-abstract class Main {
-    public static void main(String[] args) {
-        Tester test1 = new Tester();
-        test1.addCircle(new Circle(3.0, new Point(0.0, 0.0)));
-        test1.addCircle(new Circle(9.0, new Point(2.0, 2.0)));
-        test1.show();
-    }
-}
