--- students/21K0597/21K0597-p02/src/main/java/ru/mirea/practice/Task4/Main.java
@@ -1,11 +1,0 @@
-package ru.mirea.practice.Task4;
-
-abstract class Main {
-    public static void main(String[] args) {
-        Shop taverna = new Shop();
-        taverna.addComputer(new Computer("computer1", "gtx 1060", "B550", "16 gb 2400 gHz", "ryzen 5600x"));
-        taverna.addComputer(new Computer("computer1", "gtx 1080", "B550", "16 gb 2400 gHz", "ryzen 5600x"));
-        System.out.print(taverna.search("gtx 1080", "B550", "16 gb 2400 gHz", "ryzen 5600x"));
-        System.out.print(taverna.search("gtx 1080", "B550", "32 gb 2400 gHz", "ryzen 5600x"));
-    }
-}
