--- students/21K0597/21K0597-p02/src/main/java/ru/mirea/practice/Task5/Dog.java
@@ -1,44 +1,0 @@
-package ru.mirea.practice.Task5;
-
-public class Dog {
-    private String name;
-    private int age;
-
-    public Dog(String name, int age) {
-        this.name = name;
-        this.age = age;
-    }
-
-    public Dog() {
-        this.name = "dog without name(";
-        this.age = 0;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-
-    @Override
-    public String toString() {
-        return "Dog{" +
-                "name='" + name + '\'' +
-                ", age=" + age +
-                '}';
-    }
-
-    public int convertAge() {
-        return this.age * 7;
-    }
-}
