--- students/21K0597/21K0597-p02/src/main/java/ru/mirea/practice/Task6/Circle.java
@@ -1,45 +1,0 @@
-package ru.mirea.practice.Task6;
-
-import java.util.Objects;
-
-public class Circle {
-    private double radius;
-
-    public Circle(double radius) {
-        this.radius = radius;
-    }
-
-    public Circle() {
-        this.radius = 1;
-    }
-
-    public double getRadius() {
-        return radius;
-    }
-
-    public void setRadius(double radius) {
-        this.radius = radius;
-    }
-
-    public double sqare() {
-        return Math.PI * this.radius * this.radius;
-    }
-
-    public double length() {
-        return Math.PI * 2 * this.radius;
-
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Circle circle = (Circle) o;
-        return Double.compare(circle.radius, radius) == 0;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(radius);
-    }
-}
