+++ students/21K0597/21K0597-p02/src/main/java/ru/mirea/practice/task3/Point.java
@@ -1,0 +1,42 @@
+package ru.mirea.practice.task3;
+
+public class Point {
+    private double x;
+    private double y;
+
+    public Point(double x, double y) {
+        this.x = x;
+        this.y = y;
+    }
+
+    public Point() {
+        this.x = 0.0;
+        this.y = 0.0;
+    }
+
+    public double getX() {
+        return x;
+    }
+
+    public double getY() {
+        return y;
+    }
+
+    public void setX(double x) {
+        this.x = x;
+    }
+
+    public void setY(double y) {
+        this.y = y;
+    }
+
+    public void setXY(double x, double y) {
+        this.x = x;
+        this.y = y;
+    }
+
+    @Override
+    public String toString() {
+        return "center{" + "x=" + x + ", y=" + y + '}';
+    }
+}
