+++ students/21K0597/21K0597-p02/src/main/java/ru/mirea/practice/task4/Computer.java
@@ -1,0 +1,70 @@
+package ru.mirea.practice.task4;
+
+public class Computer {
+    String name;
+    String videoCard;
+    String motherBoard;
+    String memory;
+    String cpu;
+
+    public Computer(String name, String videoCard, String motherBoard, String memory, String cpu) {
+        this.name = name;
+        this.videoCard = videoCard;
+        this.motherBoard = motherBoard;
+        this.memory = memory;
+        this.cpu = cpu;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getVideoCard() {
+        return videoCard;
+    }
+
+    public void setVideoCard(String videoCard) {
+        this.videoCard = videoCard;
+    }
+
+    public String getMotherBoard() {
+        return motherBoard;
+    }
+
+    public void setMotherBoard(String motherBoard) {
+        this.motherBoard = motherBoard;
+    }
+
+    public String getMemory() {
+        return memory;
+    }
+
+    public void setMemory(String memory) {
+        this.memory = memory;
+    }
+
+    public String getCpu() {
+        return cpu;
+    }
+
+    public void setCpu(String cpu) {
+        this.cpu = cpu;
+    }
+
+    @Override
+    public String toString() {
+        return "Computer{" +
+                "name='" + name + '\'' +
+                ", videoCard='" + videoCard + '\'' +
+                ", motherBoard='" + motherBoard + '\'' +
+                ", memory='" + memory + '\'' +
+                ", cpu='" + cpu + '\'' +
+                '}';
+    }
+
+
+}
