+++ students/21K0597/21K0597-p02/src/main/java/ru/mirea/practice/task7/Tester.java
@@ -1,0 +1,14 @@
+package ru.mirea.practice.task7;
+
+abstract class Tester {
+    public static void main(String[] args) {
+        Bookshelf poloshka = new Bookshelf();
+        poloshka.addBook(new Book("Xia", "book without name", 2000));
+        poloshka.addBook(new Book("Xia", "book with name", 1920));
+        poloshka.addBook(new Book("Xia", "book with name Part 2", 1921));
+        poloshka.bookSort();
+        System.out.print(poloshka.getOldest().toString() + "\n");
+        System.out.print(poloshka.getYangest().toString() + "\n");
+        poloshka.show();
+    }
+}
